Restructure repo to mc
