+++ Updater/source/__init__.py
@@ -1,9 +1,9 @@
 """Package for the Updater module."""
 
 from source import path
-from source import download
+from Updater.source.mc import remote
 from source import exceptions
 from source import creds
-from source import pmc
+from source import mc
 from source import gui
 from source import utils
