+++ Updater/source/mc/__init__.py
@@ -1,0 +1,4 @@
+"""The PMC package for Hominum."""
+
+from source.mc.main import MCManager
+from source.mc.authentication import AuthenticationHandler
