+++ Updater/source/mc/main.py
@@ -5,18 +5,24 @@ from portablemc.fabric import FabricVersion
 
 
 class MCManager:
-    """MCManager is a class that handles PortableMC."""
+    """MCManager is a class that handles Minecraft"""
     def __init__(self, context: Context):
         self.context = context
 
-    def provision_version(self, version: str) -> FabricVersion:
+    def provision_version(self, vanilla_version: str, loader_version: str=None) -> FabricVersion:
         """
         Provisions a version for PortableMC.
+
+        Parameters:
+        - vanilla_version (str): The fabric version.
+        - loader_version (str): The loader version. Defaults to latest.
 
         Returns:
         - FabricVersion: The version for PortableMC.
         """
-        return FabricVersion.with_fabric(version, context=self.context)
+        return FabricVersion.with_fabric(
+            vanilla_version=vanilla_version, loader_version=loader_version, context=self.context
+        )
 
     def provision_environment(self, version: FabricVersion, watcher: Watcher=None) -> Environment:
         """
