--- Updater/source/pmc/__init__.py
@@ -1,4 +1,0 @@
-"""The PMC package for Hominum."""
-
-from source.pmc.main import MCManager
-from source.pmc.authentication import AuthenticationHandler
